Use separate temp dirs for main & benchmark builds
Fixes #6324
